feat(discovery): implement title search with validated pagination

+++ src/main/java/com/ott/streaming/dto/discovery/PaginationInfoPayload.java
@@ -3,7 +3,7 @@ package com.ott.streaming.dto.discovery;
 public record PaginationInfoPayload(
         int page,
         int size,
-        long totalElements,
+        int totalElements,
         int totalPages,
         boolean hasNext,
         boolean hasPrevious
